+++ NEWS.md
@@ -12,6 +12,7 @@ New features(Analysis):
   Phan's analysis of compound conditions and assignments/negations in conditions has improved enough that it should no longer be necessary.
 + Import more specific phpdoc/real array return types for internal global functions from opcache.
 + Emit `PhanUndeclaredVariable` and other warnings about arguments when there are too many parameters for methods. (#3245)
++ Infer real types of array/iterable keys and values in more cases.
 
 Maintenance:
 + Make `\Phan\Library\None` a singleton in internal uses.